requirements.txt
# Google Cloud & AI Platform
google-cloud-firestore>=2.15.0
google-cloud-aiplatform>=1.43.0
google-genai>=1.2.0
google-adk>=1.3.0
vertexai>=1.43.0

# Vector Database (Latest Pinecone)
pinecone>=7.0.0

# Data Processing
numpy>=1.24.0
scikit-learn>=1.4.0
pandas>=2.0.0

# API & Web Framework
pydantic>=2.5.0
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# Development Tools
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0